Generating a sample of points on a pipe

Points are drawn from a uniform distribution between -1 and 1, the pipe structure is generated by rejecting points if they are not on a circle with radius 1 and thickness t in the last two parameters.

pipeData(n, p, t = 0.1)

Arguments

n
sample dimensionality

p
number of sample points to generate

t
thickness of circle, default=0.1

Value

sample points in tibble format

Examples

pipeData(4, 100)

#> New names:
#> * `` -> ...1
#> * `` -> ...2
#> * `` -> ...3
#> * `` -> ...4

#> # A tibble: 100 x 4
#>       ...1   ...2   ...3    ...4
#>      <dbl>  <dbl>  <dbl>   <dbl>
#>  1 -0.135   0.809  0.974  0.0940
#>  2 -0.0383  0.559 -0.957 -0.164 
#>  3 -0.815  -0.410  0.376  0.889 
#>  4 -0.175   0.248  0.470  0.844 
#>  5 -0.928  -0.615 -0.166  0.956 
#>  6  0.216  -0.595  0.504 -0.828 
#>  7  0.323  -0.460 -0.919 -0.298 
#>  8  0.0522 -0.756  0.984 -0.187 
#>  9 -0.403   0.526  0.981  0.348 
#> 10 -0.163  -0.747 -0.412  0.914 
#> # ... with 90 more rows

pipeData(2, 100, 0.5)

#> New names:
#> * `` -> ...1
#> * `` -> ...2

#> # A tibble: 100 x 2
#>       ...1    ...2
#>      <dbl>   <dbl>
#>  1  0.259   0.912 
#>  2  0.481  -0.574 
#>  3  0.444  -0.943 
#>  4  0.0535  0.924 
#>  5  0.718   0.168 
#>  6  0.898   0.0803
#>  7 -0.739   0.434 
#>  8  0.224  -0.914 
#>  9 -0.807   0.807 
#> 10  0.831   0.0110
#> # ... with 90 more rows