Generating a sample of points on a pipe

Points are drawn from a uniform distribution between -1 and 1, the pipe structure is generated by rejecting points if they are not on a circle with radius 1 and thickness t in the last two parameters.

Usage

pipe_data(n, p, t = 0.1)

Arguments

n

number of sample points to generate

p

sample dimensionality

t

thickness of circle, default=0.1

Value

sample points in tibble format

Examples

pipe_data(100, 4)
#> # A tibble: 100 × 4
#>         V1       V2     V3      V4
#>      <dbl>    <dbl>  <dbl>   <dbl>
#>  1 -0.0105  0.408    0.381  0.921 
#>  2  0.807  -0.699   -0.948  0.126 
#>  3  0.426   0.980   -0.859 -0.410 
#>  4 -0.687  -0.00813  0.257 -0.945 
#>  5  0.452   0.584    0.705 -0.769 
#>  6 -0.745  -0.885   -0.212  0.952 
#>  7  0.583  -0.287    0.110  0.951 
#>  8  0.281  -0.993    0.523  0.822 
#>  9  0.967  -0.0456   0.954 -0.0628
#> 10  0.941   0.292    0.757  0.590 
#> # ℹ 90 more rows
pipe_data(100, 2, 0.5)
#> # A tibble: 100 × 2
#>        V1      V2
#>     <dbl>   <dbl>
#>  1  0.834  0.0861
#>  2  0.838 -0.740 
#>  3 -0.922  0.556 
#>  4  0.607  0.393 
#>  5  0.991  0.199 
#>  6  0.619 -0.913 
#>  7  0.725  0.454 
#>  8 -0.642 -0.438 
#>  9 -0.586  0.755 
#> 10 -0.356 -0.943 
#> # ℹ 90 more rows