Generating spiral sample

n-2 points are drawn from a normal distribution with mean=0, sd=1, the points in the final two direction are sampled along a spiral by samping the angle from a normal distribution with mean=0, sd=2*pi (absolute values are used to fix the orientation of the spiral).

spiralData(n, p)

Arguments

n

sample dimensionality

p

number of sample points to generate

Value

sample points in matrix format

Examples

spiralData(4, 100)
#>            [,1]         [,2]         [,3]          [,4]
#> v -1.1841315551  1.496071761 -0.183050936 -0.5042149431
#> v -0.8766481966 -1.339466676 -0.679367881 -0.9097388028
#> v  1.3714812221 -1.069745688 -0.114705800 -0.5381493979
#> v -0.9952898446 -0.709127173  0.120066435  0.1442108963
#> v -0.8356335587  0.471752612 -0.609869007 -1.6862767184
#> v  0.0348125366  0.362560413 -0.152708542  0.2684060368
#> v  0.3436103670  0.273656710 -0.076875045  0.2738610043
#> v -1.0016031875 -1.118568138 -1.038222451 -0.2369328926
#> v -1.3034629061 -0.437998449  1.226689768 -0.4847673568
#> v  0.3397608843 -0.203657727  0.136128034  0.8589610055
#> v  1.7463405784  0.482781860  0.040277322  0.2368337347
#> v -0.9472438248  0.039435646 -0.396268232 -0.2553118429
#> v  0.9021147197  0.358692081 -0.361385356 -0.3276534176
#> v -0.0003015099  1.116019831  0.123349743  0.1354848867
#> v  0.8881343882 -0.055046300  0.503271881  0.6464387560
#> v  0.0310866508 -1.522310495 -0.425371906 -0.0820153139
#> v  0.0102627588 -1.585926944  0.483434555  0.6655238812
#> v -0.1454110068  1.617246199  0.122047146  0.1391210963
#> v  0.4405500124  0.638182647  0.668738660  0.4073334809
#> v  0.3246503372 -2.020733876 -0.956723401 -0.5283871766
#> v -0.2004494320 -1.131138060  0.136227639 -0.5785421473
#> v  1.7628243484 -0.316149358 -0.251458236 -0.4561361512
#> v  1.8559522753 -0.436650654 -0.469878518 -1.0581293986
#> v -0.2282912131  0.834798318 -0.344191167 -0.3546238163
#> v -0.1228996866  0.635431456  0.643592082  0.4585173437
#> v -0.2414560094 -2.052205266  0.331152750 -0.5322890417
#> v -0.8612852256  0.158706115  0.119028844  0.1467023623
#> v  0.5491974837 -1.254691014 -0.292384250 -0.4181281514
#> v -0.1708013774 -1.712822869  0.132172734  0.0856596927
#> v -0.9396035748  0.831902588 -0.195368050  0.2562383142
#> v  0.8120804718 -0.050536012  0.901558301 -0.9056216135
#> v  1.5194423995  1.900290818  0.212329008 -0.5686268493
#> v -1.3963366453 -0.974842143 -0.342631411  0.1482417621
#> v -0.7504765806  0.180286558 -0.378383715 -0.2963024412
#> v -0.7701020618  2.647191487 -0.130351664 -0.5314323287
#> v -0.1607256448  1.618261961 -0.372514734  0.1043361271
#> v  1.7093481157 -2.390214581  0.539917377 -0.3896063446
#> v -0.5457498903  0.456843249 -0.125697469 -0.5334916660
#> v -0.5702753839  0.653609001 -0.420414989 -0.0312257672
#> v -0.1325411821  0.229463115 -0.330229722  0.1630598049
#> v -1.1487531310 -0.626067514  0.188076942 -0.5728773030
#> v -0.9736653240 -0.710160047  0.110648040  0.0126314851
#> v -1.4450643822  0.938729725  0.117118992  0.0229438418
#> v  0.8683819683 -0.586856447  0.041445832  0.2360914093
#> v -0.1392055960  0.676660160  0.648161453  0.4498866351
#> v -0.0054017584  1.230943365  0.127573157  0.1212420407
#> v -0.5405283705  0.963181072 -0.016747234 -0.5680447077
#> v -2.4370368153  1.269252594 -0.107721522  0.2741168953
#> v -0.3640824410 -0.977691835 -0.282170566 -0.4284227279
#> v -0.7918942131  1.795122053  0.979287759  1.0541865507
#> v -0.6288573969  0.547835807  0.679213431  0.3827535683
#> v  0.0598843448  0.504131970  0.138151044 -0.5784146093
#> v -1.5886029787  1.122436600  0.336514665 -0.5300027527
#> v  0.0746456100 -2.800834197 -0.348124540 -1.1163049538
#> v  0.9062143589  0.385858224  0.076758510  0.8732638699
#> v -1.7938061176  0.158402702 -0.424051688 -0.1345287154
#> v -0.0802218062 -0.622310944 -0.100112903 -0.5438992217
#> v  0.4321533407  0.995597638 -0.424760575 -0.0699633640
#> v -0.4793830473 -0.914601374  0.113522892  0.0168938522
#> v  0.9112071833  0.116916346 -0.292195977  0.2000603795
#> v  0.7285455061 -0.469120419 -0.395549782 -1.0957021911
#> v  0.2851892396 -0.595438199  0.724909936 -0.0222430360
#> v -0.1720538391  0.535371046 -0.376989954  0.0964095111
#> v -0.8239034027  0.104849405  0.100762529  0.0007713331
#> v -0.4515997490  0.800731333  0.018168144  0.2491385659
#> v  0.0339915266 -0.261726400  0.002040647  0.2562752537
#> v -2.4000573695 -0.314748841 -0.185971799  0.2595031212
#> v -0.7954683511  0.649933558  0.020718126  0.2478764387
#> v  1.8094944825  0.051080520 -0.239680174 -0.4656365284
#> v -1.3936313039  1.165754571 -0.921756709 -0.6007377851
#> v  0.0393348066  0.287345287 -0.272782973 -0.4373825274
#> v -0.9298408182 -1.041219881  0.664590808 -0.2116695419
#> v  0.9293446286  1.109428537 -0.404981007  0.0320948161
#> v  0.3240792966 -0.335717843 -0.271288722  0.2161665891
#> v  0.7347498059  0.932446531  0.622767618 -0.2859414475
#> v -1.7820720979 -0.685050272  0.649729798 -0.2406130994
#> v -0.8077797735 -0.448933641 -0.890377324  0.4463810924
#> v -0.4512672987 -0.174984512 -0.773500514  0.5964975298
#> v  0.7535878778  0.559350774  1.232730111  0.6781016013
#> v  0.2924207236 -2.057669843 -0.113245367 -0.5387468629
#> v -0.6316115455  1.265117270  0.115152672  0.1551832059
#> v -1.6832779859  1.188388111 -0.971083081  0.2899914741
#> v  2.2468224487  0.693021578  0.449298615  0.6955496642
#> v -0.1499828731  0.775419569  1.358729775  0.0343352168
#> v -1.1027979956  0.466849407 -0.080445205  0.2740676824
#> v -2.0194075812  1.076187135 -0.423129951 -0.0513362689
#> v  0.4665303640 -1.552105387  0.129751022  0.0583780338
#> v -1.2451144993  0.670415548 -1.021304691  0.1312181854
#> v -0.3508083048 -0.110139767  0.734856157  0.0814080294
#> v -0.0706885084 -0.007169724 -0.095596735  0.2744261693
#> v  0.8993717560 -0.815246072  1.017987696 -0.7930758886
#> v  0.4146715143 -1.767657026 -0.398802857 -0.2484833674
#> v  0.8998866011 -0.700016699  0.151904593 -0.5773176005
#> v  0.7174210815 -0.641243953 -0.184111582  0.2601094345
#> v -0.0891892429  0.434454199  0.083043122  0.2015624896
#> v -0.9097624692 -0.879968972 -0.965243137  0.3039574772
#> v  0.4539588422 -1.038688286 -0.393669651 -0.2619905512
#> v  0.6737462463  0.593841192  0.734609464  0.0731223896
#> v  0.0424708909  0.071963207 -0.069247147  0.2732607621
#> v  0.3316569697  0.733984440 -0.008559944  0.2602271121