Estimating squint angle of 2-d structure in high-d dataset under selected index.

We estimate the squint angle by interpolating from a random starting plane towards the optimal view until the index value of the selected index function is above the selected cutoff. Since this depends on the direction, this is repeated with n randomly selected planes giving a distribution representative of the squint angle.

squintAngleEstimate(
  data,
  indexF,
  cutoff,
  structurePlane,
  n = 100,
  stepSize = 0.01
)

Arguments

data
Input data

indexF
Index function

cutoff
Threshold index value above which we assume the structure to be visible

structurePlane
Plane defining the optimal view

n
Number of random starting planes (default = 100)

stepSize
Interpolation step size fixing the accuracy (default = 0.01)

Value

numeric vector containing all squint angle estimates

Examples

# \donttest{
data <- spiralData(4, 100)
indexF <- scagIndex("Skinny")
cutoff <- 0.7
structurePlane <- basisMatrix(3,4,4)
squintAngleEstimate(data, indexF, cutoff, structurePlane, n=1)

#> target_dist - cur_dist: 0 
#> generation:  dist =   1.277016 
#> target_dist - cur_dist: 1.277016 
#> generation:  dist =   1.406553 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.697809 
#> target_dist - cur_dist: 0 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.697809 
#> target_dist - cur_dist: 1.697809 
#> target_dist - cur_dist: 1.687809 
#> target_dist - cur_dist: 1.677809 
#> target_dist - cur_dist: 1.667809 
#> target_dist - cur_dist: 1.657809 
#> target_dist - cur_dist: 1.647809 
#> target_dist - cur_dist: 1.637809 
#> target_dist - cur_dist: 1.627809 
#> target_dist - cur_dist: 1.617809 
#> target_dist - cur_dist: 1.607809 
#> target_dist - cur_dist: 1.597809 
#> target_dist - cur_dist: 1.587809 
#> target_dist - cur_dist: 1.577809 
#> target_dist - cur_dist: 1.567809 
#> target_dist - cur_dist: 1.557809 
#> target_dist - cur_dist: 1.547809 
#> target_dist - cur_dist: 1.537809 
#> target_dist - cur_dist: 1.527809 
#> target_dist - cur_dist: 1.517809 
#> target_dist - cur_dist: 1.507809 
#> target_dist - cur_dist: 1.497809 
#> target_dist - cur_dist: 1.487809 
#> target_dist - cur_dist: 1.477809 
#> target_dist - cur_dist: 1.467809 
#> target_dist - cur_dist: 1.457809 
#> target_dist - cur_dist: 1.447809 
#> target_dist - cur_dist: 1.437809 
#> target_dist - cur_dist: 1.427809 
#> target_dist - cur_dist: 1.417809 
#> target_dist - cur_dist: 1.407809 
#> target_dist - cur_dist: 1.397809 
#> target_dist - cur_dist: 1.387809 
#> target_dist - cur_dist: 1.377809 
#> target_dist - cur_dist: 1.367809 
#> target_dist - cur_dist: 1.357809 
#> target_dist - cur_dist: 1.347809 
#> target_dist - cur_dist: 1.337809 
#> target_dist - cur_dist: 1.327809 
#> target_dist - cur_dist: 1.317809 
#> target_dist - cur_dist: 1.307809 
#> target_dist - cur_dist: 1.297809 
#> target_dist - cur_dist: 1.287809 
#> target_dist - cur_dist: 1.277809 
#> target_dist - cur_dist: 1.267809 
#> target_dist - cur_dist: 1.257809 
#> target_dist - cur_dist: 1.247809 
#> target_dist - cur_dist: 1.237809 
#> target_dist - cur_dist: 1.227809 
#> target_dist - cur_dist: 1.217809 
#> target_dist - cur_dist: 1.207809 
#> target_dist - cur_dist: 1.197809 
#> target_dist - cur_dist: 1.187809 
#> target_dist - cur_dist: 1.177809 
#> target_dist - cur_dist: 1.167809 
#> target_dist - cur_dist: 1.157809 
#> target_dist - cur_dist: 1.147809 
#> target_dist - cur_dist: 1.137809 
#> target_dist - cur_dist: 1.127809 
#> target_dist - cur_dist: 1.117809 
#> target_dist - cur_dist: 1.107809 
#> target_dist - cur_dist: 1.097809 
#> target_dist - cur_dist: 1.087809 
#> target_dist - cur_dist: 1.077809 
#> target_dist - cur_dist: 1.067809 
#> target_dist - cur_dist: 1.057809 
#> target_dist - cur_dist: 1.047809 
#> target_dist - cur_dist: 1.037809 
#> target_dist - cur_dist: 1.027809 
#> target_dist - cur_dist: 1.017809 
#> target_dist - cur_dist: 1.007809 
#> target_dist - cur_dist: 0.997809 
#> target_dist - cur_dist: 0.987809 
#> target_dist - cur_dist: 0.977809 
#> target_dist - cur_dist: 0.967809 
#> target_dist - cur_dist: 0.957809 
#> target_dist - cur_dist: 0.947809 
#> target_dist - cur_dist: 0.937809 
#> target_dist - cur_dist: 0.927809 
#> target_dist - cur_dist: 0.917809 
#> target_dist - cur_dist: 0.907809 
#> target_dist - cur_dist: 0.897809 
#> target_dist - cur_dist: 0.887809 
#> target_dist - cur_dist: 0.877809 
#> target_dist - cur_dist: 0.867809 
#> target_dist - cur_dist: 0.857809 
#> target_dist - cur_dist: 0.847809 
#> target_dist - cur_dist: 0.837809 
#> target_dist - cur_dist: 0.827809 
#> target_dist - cur_dist: 0.817809 
#> target_dist - cur_dist: 0.807809 
#> target_dist - cur_dist: 0.797809 
#> target_dist - cur_dist: 0.787809 
#> target_dist - cur_dist: 0.777809 
#> target_dist - cur_dist: 0.767809 
#> target_dist - cur_dist: 0.757809 
#> target_dist - cur_dist: 0.747809 
#> target_dist - cur_dist: 0.737809 
#> target_dist - cur_dist: 0.727809 
#> target_dist - cur_dist: 0.717809 
#> target_dist - cur_dist: 0.707809 
#> target_dist - cur_dist: 0.697809 
#> target_dist - cur_dist: 0.687809 
#> target_dist - cur_dist: 0.677809 
#> target_dist - cur_dist: 0.667809 
#> target_dist - cur_dist: 0.657809 
#> target_dist - cur_dist: 0.647809 
#> target_dist - cur_dist: 0.637809 
#> target_dist - cur_dist: 0.627809 
#> target_dist - cur_dist: 0.617809 
#> target_dist - cur_dist: 0.607809 
#> target_dist - cur_dist: 0.597809 
#> target_dist - cur_dist: 0.587809 
#> target_dist - cur_dist: 0.577809 
#> target_dist - cur_dist: 0.567809 
#> target_dist - cur_dist: 0.557809 
#> target_dist - cur_dist: 0.547809 
#> target_dist - cur_dist: 0.537809 
#> target_dist - cur_dist: 0.527809 
#> target_dist - cur_dist: 0.517809 
#> target_dist - cur_dist: 0.507809 
#> target_dist - cur_dist: 0.497809 
#> target_dist - cur_dist: 0.487809 
#> target_dist - cur_dist: 0.477809 
#> target_dist - cur_dist: 0.467809 
#> target_dist - cur_dist: 0.457809 
#> target_dist - cur_dist: 0.447809 
#> target_dist - cur_dist: 0.437809 
#> target_dist - cur_dist: 0.427809 
#> target_dist - cur_dist: 0.417809 
#> target_dist - cur_dist: 0.407809 
#> target_dist - cur_dist: 0.397809 
#> target_dist - cur_dist: 0.387809 
#> target_dist - cur_dist: 0.377809 
#> target_dist - cur_dist: 0.367809 
#> target_dist - cur_dist: 0.357809 
#> target_dist - cur_dist: 0.347809 
#> target_dist - cur_dist: 0.337809 
#> target_dist - cur_dist: 0.327809 
#> target_dist - cur_dist: 0.317809 
#> target_dist - cur_dist: 0.307809 
#> target_dist - cur_dist: 0.297809 
#> target_dist - cur_dist: 0.287809 
#> target_dist - cur_dist: 0.277809 
#> target_dist - cur_dist: 0.267809 
#> target_dist - cur_dist: 0.257809 
#> target_dist - cur_dist: 0.247809 
#> target_dist - cur_dist: 0.237809 
#> target_dist - cur_dist: 0.227809 
#> target_dist - cur_dist: 0.217809 
#> target_dist - cur_dist: 0.207809 
#> target_dist - cur_dist: 0.197809 
#> target_dist - cur_dist: 0.187809 
#> target_dist - cur_dist: 0.177809 
#> target_dist - cur_dist: 0.167809 
#> target_dist - cur_dist: 0.157809 
#> target_dist - cur_dist: 0.147809 
#> target_dist - cur_dist: 0.137809 
#> target_dist - cur_dist: 0.127809 
#> target_dist - cur_dist: 0.117809 
#> target_dist - cur_dist: 0.107809 
#> target_dist - cur_dist: 0.09780896 
#> target_dist - cur_dist: 0.08780896 
#> target_dist - cur_dist: 0.07780896 
#> target_dist - cur_dist: 0.06780896 
#> target_dist - cur_dist: 0.05780896 
#> target_dist - cur_dist: 0.04780896 
#> target_dist - cur_dist: 0.03780896 
#> target_dist - cur_dist: 0.02780896 
#> target_dist - cur_dist: 0.01780896 
#> target_dist - cur_dist: 0.007808965 

#> [1] 0.07660425

# }