Time each index evaluation for projections in the tour path.

Index evaluation timing may depend on the data distribution, we evaluate the computing time for a set of different projections to get an overview of the distribution of computing times.

timeSequence(d, t, idx, pmax)

Arguments

d

Input data in matrix format

t

List of projection matrices (e.g. interpolated tour path)

idx

Index function

pmax

Maximum number of projections to evaluate (cut t if longer than pmax)

Value

numeric vector containing all distances

Examples

d <- spiralData(4, 500)
t <- purrr::map(1:10, ~ tourr::basis_random(4))
idx <- scagIndex("stringy")
timeSequence(d, t, idx, 10)
#>        t  i
#> 1  0.053  1
#> 2  0.055  2
#> 3  0.055  3
#> 4  0.052  4
#> 5  0.056  5
#> 6  0.056  6
#> 7  0.054  7
#> 8  0.058  8
#> 9  0.071  9
#> 10 0.055 10