Generating a sample of points on a pipe

Points are drawn from a uniform distribution between -1 and 1, the pipe structure is generated by rejecting points if they are not on a circle with radius 1 and thickness t in the last two parameters.

pipeData(n, p, t = 0.1)

Arguments

n
sample dimensionality

p
number of sample points to generate

t
thickness of circle, default=0.1

Value

sample points in tibble format

Examples

pipeData(4, 100)

#> # A tibble: 100 x 4
#>         V1     V2     V3     V4
#>      <dbl>  <dbl>  <dbl>  <dbl>
#>  1 -0.0160 -0.428 -0.983  0.102
#>  2  0.626   0.287 -0.862  0.457
#>  3 -0.146   0.539  0.424 -0.873
#>  4 -0.719   0.421 -0.747  0.671
#>  5  0.594   0.439  0.456 -0.904
#>  6  0.262  -0.620 -0.593  0.857
#>  7  0.280   0.765  0.892  0.419
#>  8  0.993   0.290  0.836  0.590
#>  9  0.423  -0.816 -0.788 -0.550
#> 10  0.527   0.688  0.410  0.897
#> # … with 90 more rows

pipeData(2, 100, 0.5)

#> # A tibble: 100 x 2
#>        V1      V2
#>     <dbl>   <dbl>
#>  1 -0.273 -0.929 
#>  2  0.743  0.0107
#>  3  0.875 -0.0537
#>  4  0.732  0.907 
#>  5 -0.959 -0.608 
#>  6  0.339 -0.829 
#>  7 -0.776  0.372 
#>  8 -0.721 -0.131 
#>  9 -0.733 -0.334 
#> 10  0.637  0.752 
#> # … with 90 more rows