Generating spiral sample

n-2 points are drawn from a normal distribution with mean=0, sd=1, the points in the final two direction are sampled along a spiral by samping the angle from a normal distribution with mean=0, sd=2*pi (absolute values are used to fix the orientation of the spiral).

spiralData(n, p)

Arguments

n
sample dimensionality

p
number of sample points to generate

Value

sample points in matrix format

Examples

spiralData(4, 100)

#>           [,1]         [,2]         [,3]         [,4]
#> v  0.699458213  0.420749052 -1.036870797  0.047088665
#> v  0.028626369 -0.307044403  0.039238513  0.237485300
#> v -1.668234902  0.913667446  0.735014504  0.108085769
#> v -0.854521406  0.829645624  0.020140939  0.248165480
#> v -0.298698402  0.058397955  0.118900164  0.026308952
#> v -0.386304571  0.553146957 -0.365186486 -0.321111640
#> v  0.432207818 -0.180889603 -0.944653361  0.349063296
#> v -0.170520042 -0.183991976 -0.368636818  0.110862157
#> v  0.618635451 -0.418076761 -0.346296614  0.143537064
#> v  0.062490408 -0.325947562  0.474935929  0.673317361
#> v -1.030284809 -1.994814416 -0.917468558  0.400997356
#> v  0.472361149  0.061306875  0.734055253  0.137753152
#> v -1.246621804 -0.431414995 -0.243672153 -0.462480439
#> v  0.688959955 -0.373306023  0.707164977  0.301003042
#> v -0.484026232 -0.301411015  0.002322317  0.256162551
#> v -0.902722153 -0.218671502 -0.881314561 -0.670757503
#> v  0.012081142 -1.462349027 -0.967599874 -0.502699982
#> v -1.052127461  0.920678052 -0.407996914  0.022610987
#> v -0.329105928  0.253252845  0.733611543  0.052890285
#> v  0.862736211 -0.095968112 -0.408494436  0.020963077
#> v  0.832358398 -0.571346695 -1.653935384 -0.357191004
#> v -1.396796640  0.901393170  0.234894715 -0.563738950
#> v  1.776680608  0.193685397 -0.228816079 -0.473909349
#> v  1.700058074 -0.599405946 -0.051394494  0.270991847
#> v  0.448238081  1.291826956 -0.880936558  0.460995017
#> v -0.030615390  0.487155920 -0.412714054  0.005796955
#> v  0.611951853 -0.464028097 -0.063444218  0.272657576
#> v  0.004356106  0.039522727 -0.905609193  0.421542810
#> v -1.379596738  0.827211992  0.129508876  0.112383478
#> v -0.581278296 -1.219654411  0.384309086 -0.506894259
#> v -0.869415763 -0.275297954 -0.077231589  0.890705324
#> v  1.143544128 -0.033595699 -0.422266042 -0.151387171
#> v  0.339072344 -0.787546867 -0.036121297 -0.563683410
#> v -1.127718046 -1.530245521 -0.640853894 -0.942636713
#> v -0.171330500  0.239694471 -0.132004392  0.271926103
#> v  2.420930320  0.696484772 -0.136583831 -1.180157965
#> v -0.450178648  1.802517472  1.328555384  0.392695115
#> v  0.044168205  0.475669363  0.734801145  0.118973316
#> v -0.359263442  0.747456094  0.121047339  0.030778263
#> v  1.053804209  1.081708091  0.127853398  0.120094988
#> v  0.274615797 -0.857221702  0.654261635 -0.232137387
#> v -1.443809924  0.594575761 -0.048910271  0.270577725
#> v -1.888178013 -0.832101980 -0.075460995  0.273766108
#> v  0.023076219  1.465416379  0.080195269 -1.203542010
#> v  1.010535190  1.402136895 -0.515852170  0.789708193
#> v -1.149220362 -1.315667269 -0.270110760 -0.439850433
#> v  0.108788571 -2.654821839 -0.353992207 -0.339737595
#> v -0.971058882 -0.172120501  0.710789165 -0.087174699
#> v -1.404287062  0.181365211 -0.355433917  0.131069581
#> v -0.353934941 -0.286801571 -0.249402466  0.878342441
#> v  0.717935654 -1.368989704  0.113486343 -0.579568942
#> v  1.174950132 -1.336277463  0.021955240  0.247250223
#> v  1.290175952 -0.700626152 -0.876837635  0.467168822
#> v -0.696091946  2.060206827 -0.111691987  0.273901643
#> v  1.483909616  0.093635749 -0.206986516  0.251725065
#> v  0.125540490  0.708059319  0.626323015 -0.280400486
#> v -1.163520553  0.548165277 -0.095224055  0.274427329
#> v -0.468856671  1.613090761 -0.400009767  0.046176712
#> v -0.002386705 -1.523147601 -0.373918729  0.101897077
#> v  0.145965117  0.391357756 -0.234443801  0.238855149
#> v -1.258791948  0.554450666 -0.272622770  0.215214663
#> v -0.620354358  1.168033410 -0.425238819  0.830162751
#> v  0.285242241 -0.607892055  0.694670798  0.341301696
#> v  1.108352952 -1.381282690 -0.715594937 -0.875907120
#> v -0.187394994  0.377211276 -0.127815197  0.272453112
#> v  0.675587614 -0.745156850 -0.231538530  0.240370087
#> v  1.179962907 -0.556589235  0.703942974 -0.111130137
#> v -1.323135698  0.199336897  0.723953130  0.226975032
#> v  0.170944375 -0.006029045 -0.117050027 -0.537179965
#> v  0.141461813 -0.610794173 -0.247845108 -0.459111769
#> v -0.983504413 -1.475404571 -0.391466340 -0.267425945
#> v -0.952604100 -0.467483822 -0.170855741  0.264053930
#> v -0.034267733 -1.509246423  0.112227355  0.160904404
#> v -0.900335360  1.294312950 -0.424467082 -0.129355697
#> v -0.150995478  1.724985599 -0.358061727  0.127270240
#> v -0.616711120  0.145275012  0.440662632 -0.472736355
#> v -0.607295712 -1.997295805 -0.425650939 -0.098764559
#> v  0.071793193  1.169234800  0.715981389 -0.066474667
#> v  1.214754360 -0.200026910 -0.379848927  0.091100168
#> v  0.028604226  0.579981950  0.132146759  0.081263446
#> v -1.457266920  0.919914630  0.059056284  0.223588688
#> v -0.116114716  0.602896795  0.562942294 -0.364818368
#> v -0.160187834 -1.117706865 -0.926606115  0.384353138
#> v -0.185140085  1.948864312 -0.287189771  0.204156539
#> v  1.183893406  0.702532412 -0.349790675  0.138899196
#> v -1.056433277 -0.839280266  0.205085173 -0.570004460
#> v  0.366846781 -0.483405917 -0.284243415  0.206494428
#> v -1.221492193  0.739113901  0.062984163 -0.578663006
#> v -0.260360640 -0.549026496  0.131642610  0.096828438
#> v -0.647624989 -0.290057877 -0.405397830  0.030831771
#> v  0.632114594  1.364653938  0.024194702  0.246093110
#> v  0.390053939 -0.868782463  0.127980869  0.119560907
#> v  2.605283966 -0.297980860 -0.288444527 -0.422169682
#> v  1.353017655 -0.009176653 -0.232254163 -1.156653406
#> v  0.406904424 -0.355068017 -0.772623227  0.597421315
#> v  1.220722544 -0.532262287  0.721154674  0.242097946
#> v  1.118541491  0.936719300 -0.309656084  0.184461329
#> v -0.261437760 -1.789888624  0.437671355 -0.474756365
#> v -1.084631359 -0.319138347 -0.196332646  0.255883852
#> v  1.335629896  0.443883836  0.093353627  0.189586412