Generating spiral sample

n-2 points are drawn from a normal distribution with mean=0, sd=1, the points in the final two direction are sampled along a spiral by samping the angle from a normal distribution with mean=0, sd=2*pi (absolute values are used to fix the orientation of the spiral).

spiralData(n, p)

Arguments

n
sample dimensionality

p
number of sample points to generate

Value

sample points in matrix format

Examples

spiralData(4, 100)

#>          [,1]         [,2]         [,3]          [,4]
#> v -0.19156194 -1.203991350 -1.014096371 -0.3619752408
#> v -0.14928970 -0.988248953 -0.286322441  0.8713089772
#> v  0.76400192 -0.764220023  0.712412186  0.2812472828
#> v  0.57440044  1.454354816  1.173630125 -0.5831481210
#> v  0.37769313  0.457817317 -0.117292095  0.2735029003
#> v -0.19633722  0.532207247 -0.265884306  0.8753978035
#> v -0.11645988 -1.791435814  0.731037563  0.1761793186
#> v -0.75173731  1.443320207 -0.055669951 -0.5585371250
#> v -1.05455799  0.030424727 -0.355893810 -0.3367049946
#> v -0.19350202  0.393606874  0.118252064  0.0250520346
#> v -0.52872257  3.071007147 -0.885926708  0.4533407031
#> v  0.33070126 -1.759479023  0.734702696  0.0759022651
#> v  1.03686204  0.034712279 -0.403344175 -1.0920728500
#> v -0.37408213 -1.258362453 -0.375083333 -0.3028489907
#> v -1.36870786 -0.942163446 -0.035972179 -0.5637197563
#> v  0.02209434  0.548882732 -0.560974276  0.7650359538
#> v -0.64446763  1.047844734 -0.130945548  0.2720651770
#> v  0.03432588  1.063948370 -0.935402815 -0.5740903028
#> v -1.58346240  0.110531327  0.005923797  0.2546852408
#> v  0.05714992 -1.083303495  0.049577486  0.2306376203
#> v -0.82368553 -1.500780397  0.127730116  0.8612546621
#> v -0.21586746  1.733784762 -0.911576385  0.4113464910
#> v -0.17647895  0.756314868  0.710117394 -0.0896737309
#> v -0.38642438  0.381516254 -0.535260191  0.7794880575
#> v  0.98374868  0.086637298 -0.387949592  0.0748395291
#> v -0.89290017  0.002229617  0.113710805  0.1580657490
#> v -0.57455455  0.235160435 -0.089067523  0.8909212174
#> v -1.39029804  0.115795724  0.515926852 -0.4129241196
#> v  0.03875288 -0.038372698  0.229669630 -0.5649516047
#> v  0.03348690  0.663318866  0.044515663  0.2340927364
#> v -0.01573655  2.087713501 -1.040093875  0.0223100357
#> v -0.30976402  1.584290933  0.493772288  0.6557392294
#> v  0.93799414 -1.176473378 -0.801290746  0.5659213510
#> v -0.17717538 -0.650604263  0.112800521  0.8651094740
#> v  0.25132590 -1.563215351  0.235826191 -0.5635176176
#> v -0.11084868 -0.761172061  0.297595955 -0.5452740228
#> v -1.05517196  0.222967963 -1.523586805  0.5900362648
#> v  2.00064170 -0.754805825 -0.301664266  0.8679184889
#> v  0.24602616  0.548893707 -0.404146203  0.0345825803
#> v -0.85724446 -0.208889627  0.112647310  0.0155473635
#> v  0.37172886 -2.300614875 -0.206722852 -0.4894038454
#> v  0.82395471  0.671560953  0.121382922  0.1408817144
#> v -0.32485542  1.165364420  0.549369758 -0.3797364702
#> v  0.98561211 -0.126323527 -0.945488081  0.3473439596
#> v  0.03215408 -2.117716447 -0.578625529 -0.9899646518
#> v  0.57567105 -0.002235961 -0.096851114 -0.5451184847
#> v  0.44867236  0.051736240 -0.020353147  0.8874578217
#> v  0.55837496  0.298702402 -0.311409596  0.1827732282
#> v -1.62537447  0.299647475  0.116556648  0.1522451424
#> v  0.33047044 -0.344358479  0.513407594 -0.4152384889
#> v -0.13629434  0.281191185  0.130145058  0.1087891614
#> v -0.03624743  0.024977216  0.123939994  0.1337459798
#> v  2.59552392  0.419283992 -0.054804534  0.2715205966
#> v  0.54486311  0.868230689  0.061801380  0.2213938664
#> v  0.07494732 -2.250901330 -1.043062689 -0.0062992838
#> v -0.48276894  0.605210639  0.188832310 -0.5727604305
#> v  0.31836219  0.103472107 -0.281393143 -0.4291824092
#> v  0.03387470 -0.649756559  0.570641346 -0.3559357896
#> v -0.04727292 -1.178515013  0.069563972  1.5075012750
#> v  1.46748561  0.552061735 -0.207176850  0.2516466539
#> v  1.83752167  0.883273304 -0.269589149 -1.1451305094
#> v  0.29372860  1.330807535  0.100642249  0.0006484856
#> v  0.94401212 -0.163110155  0.112177051  0.0148422513
#> v  0.55105923  0.754814726  0.734893353  0.0830247316
#> v -0.32052555 -0.457239570  0.715262537  1.2757955160
#> v  0.06640965  1.513114967  0.011789187  0.2521317625
#> v  0.19609740 -0.656023781 -0.396860718  0.0543079887
#> v  0.56887874 -2.102626810  0.102052430  0.0021172700
#> v  0.41691801  0.473867599  0.128772303  0.0537426601
#> v -1.63621935  1.168229652 -0.299655258  0.1936552107
#> v -0.09411706  0.467589309  0.292682535 -0.5469885690
#> v  0.39681323  0.244475509 -0.344080601 -0.3547851288
#> v -0.80520647 -1.932053070 -0.113061243 -0.5388218364
#> v  1.79248656  1.123194865 -0.856704529 -0.7082534166
#> v  1.73279950  1.275703717 -0.325432372  0.1683792770
#> v -0.13777813 -0.597937324 -0.096657884 -0.5451899703
#> v  0.07276248  1.473362511 -0.120693198  0.2732064821
#> v -1.41472141 -1.986179040 -0.149202569 -0.5225420158
#> v -0.79042763  0.758480158  1.238832022  0.6646797603
#> v  1.91913490  0.075529822 -0.029855707  0.2665665228
#> v  0.43629074  1.209748781 -0.017034268  0.8871545065
#> v  1.76502863 -0.159322027  0.535040387 -0.3945419346
#> v -0.70499533  1.126896570  1.323094366 -0.2153374449
#> v -1.24477283  1.031250388 -0.407742603 -0.2211552937
#> v -1.03789001  1.561622325  0.131532463  0.0980607899
#> v  2.32423830  2.325851550 -0.244379229  0.8791777065
#> v -0.04501885  0.195694234 -0.399009098  0.0488195257
#> v  2.01734246  1.195608651 -0.283878949 -0.4267417536
#> v  1.02629785  0.520192659  0.584032642  0.5527100842
#> v  0.39989766 -0.984527658 -0.415715257 -0.0066930434
#> v  0.98714844  2.191481010  0.103644762  0.1753598547
#> v -0.68604080  0.941399410  0.104420952  0.1741670467
#> v -1.30233941 -0.723484004 -0.750377701  0.6200620215
#> v  0.68184063  0.468260280 -0.320400532  0.1737346174
#> v -0.80033126 -0.488457510 -0.425559044 -0.1064322364
#> v  1.43517102 -0.261838719 -0.864562190  0.4850796091
#> v -0.51379272  0.772301225 -0.807562350  0.5586481494
#> v -0.01580491 -1.023289556  1.064173513  0.9548677637
#> v  0.15037144 -0.520871198  0.554318078 -0.3744028369
#> v -0.74998690 -0.786576021 -0.420272950 -0.1654268408