Generating spiral sample

n-2 points are drawn from a normal distribution with mean=0, sd=1, the points in the final two direction are sampled along a spiral by samping the angle from a normal distribution with mean=0, sd=2*pi (absolute values are used to fix the orientation of the spiral).

spiralData(n, p)

Arguments

n

sample dimensionality

p

number of sample points to generate

Value

sample points in matrix format

Examples

spiralData(4, 100)
#>           [,1]          [,2]          [,3]         [,4]
#> v -1.168696300 -1.4194533339 -0.3076103643 -0.401604162
#> v -0.427132789  1.1898281468  0.1198847990  0.144654970
#> v  0.189395976  0.0575844052 -0.4002595207 -0.244401979
#> v -0.241414580  0.7638003907 -0.2296618305 -0.473281469
#> v  0.836994139  0.0419394664  0.7092822849 -0.092730837
#> v  1.496095654  1.3811671565  0.0933454747  0.189596675
#> v -0.604558886 -0.0064125536  0.7336799974  0.144345765
#> v -0.596619301  1.2300053344  1.0132560076  1.016622338
#> v  0.546541285  1.0114553407 -0.6439816459  0.710639980
#> v  0.707837970 -0.1225158540  0.1839880311 -0.573492726
#> v -0.179360004  0.1094528569  0.0293024689  0.243337495
#> v -1.269663162  0.6205258113 -0.3152458035 -0.392724481
#> v  0.662357814  0.2458842772  0.3651124585 -0.516780474
#> v -0.434767411  0.8831759490 -0.4011778179 -0.241764623
#> v -0.256692041  1.2201357048 -0.7166109786  1.387239710
#> v -1.307042336  0.0002714681  0.5113153011  0.638320915
#> v -0.221940135 -0.2083145925 -0.0002773791 -0.571190269
#> v  0.117605618  0.0990212037 -0.6168826233  0.729779203
#> v -0.838988239 -1.0649372190  1.0817625631  0.931793300
#> v  0.510667187  1.2002630345  0.7262155071  0.213263421
#> v  0.720443607  0.5029684073  0.6737792531 -0.191895002
#> v  0.492128056 -0.1559494684 -0.3396728237  0.151926036
#> v -1.223420350 -0.2903433729 -0.1685145390  0.264682854
#> v  0.234552056  0.0118239369 -0.3797397036 -0.293537226
#> v  0.737509763 -0.8981034882  0.0980927892  0.183370008
#> v  0.913234789 -1.1051534547 -0.2722660027  0.215470203
#> v -1.369842671  0.7394716153 -0.3233485950 -0.382814819
#> v -2.280353862 -2.0112246984  1.2712790284  0.585976230
#> v -1.315052557 -1.1097503996  0.1265710627  0.045475626
#> v -0.862587391  0.9307754912 -0.4137562651 -0.198428339
#> v  1.234831926  0.4620031264 -0.0556040854 -0.558555751
#> v  1.323360624  0.4572947704 -0.0477442703 -0.560715491
#> v  2.143612831 -0.3381593680  0.1304529694  0.062360839
#> v -0.088210020  0.3550907169  0.2400866331  0.822686678
#> v  1.456000504 -0.0004656092  0.1285346623  0.117140226
#> v  0.184148214 -0.8612007223  0.1079813559 -0.579684028
#> v -0.454064172 -2.0286071844  0.1210539941  0.141732631
#> v -0.653165249  0.7200350979  0.2182394050 -0.567434190
#> v  0.962452130  0.1031380723 -0.2068115495 -0.489344993
#> v  1.095198934  0.1818922153  0.0732904731  0.211353750
#> v -0.848061384 -1.8774430764 -1.4783768785 -0.888680233
#> v  0.108067227 -1.8383419463  0.1268955056  0.123878380
#> v  0.147487879 -1.4306464098  0.1309988591  0.102922069
#> v  0.503181909  2.5126804489 -0.3520231322  0.135853580
#> v -2.178999129 -0.6781489679 -0.0739368541 -0.553027652
#> v -1.243432737 -0.5882719769  0.6183602581  0.502063559
#> v -0.502379988  0.3472986863 -0.0139538111 -1.198311785
#> v  0.462855341  0.7678697074 -0.3653668370  0.116140960
#> v -0.352577225  0.3664544857 -0.4055663444  0.030317059
#> v  0.873390014  1.7987806220  0.7218427804  0.238534892
#> v  2.175230640 -0.1835628629 -0.2055851167 -0.490156445
#> v  0.025292184  0.7770312224 -0.5985838510  0.741923605
#> v  0.311945424  0.9911976849 -0.2505742772  0.229741705
#> v -0.880261515  0.0230683213 -0.1408719990 -0.526581255
#> v  1.471071167  1.4231851069 -0.0640226258 -0.556103401
#> v -1.227862318 -0.8465572264 -0.7130354007 -0.878397008
#> v -2.536982727  1.5149668186 -0.5284014361  0.783166088
#> v  0.342217995 -0.3246704812  0.6645807624 -0.211690297
#> v  0.122596323  2.2095975457 -0.3112171691  0.865666876
#> v -0.525514066  1.1516359225  0.2851004704 -0.549543031
#> v  1.156244832  0.4109726943  0.1269179651  0.046640169
#> v  0.780177135 -1.9231777839 -0.0402600012  0.268942329
#> v  0.603717732 -1.3173483134  0.6733212978 -0.192920354
#> v  0.012853538  0.2914360940  1.3595814271  0.144097468
#> v -1.596488404  0.7905133025  0.1268897808  0.046543977
#> v  2.784656538  0.3954117043 -0.5765529350 -1.700075635
#> v  1.056610133  1.7818693119 -0.3884207432  0.843413193
#> v -0.724083154 -0.2632911657 -0.1852261385  0.259747735
#> v -0.213251900  0.7486664662  1.9870021296  0.056818810
#> v -0.205710195 -0.1435167183 -0.3451952240  0.856772750
#> v  0.764573261  0.5644685721  0.0950312719  0.187444140
#> v -0.002574686  0.5887499380  0.0950350492  0.187439247
#> v  1.396922971  0.8254166479 -0.4222863461 -0.151225214
#> v  0.545377983 -0.5363845039  0.6679222538 -1.063262143
#> v -0.712342194 -0.0945664557  0.0110850375  0.252448181
#> v  1.020446106  0.9232619914 -0.0219499551  0.264451454
#> v -0.687586188  0.2208611800 -0.4225073244 -0.149429829
#> v  0.551846388 -0.1618329852 -0.5646235811  0.762898144
#> v -0.005333142 -2.4051334407  0.7204362126 -0.046192449
#> v -1.233031680  0.0274105624 -0.2677341812  0.218655414
#> v  2.401810177  0.5541959989 -0.4197637517 -0.168568153
#> v -1.430788808  0.2565309367 -0.3606395519  0.123441647
#> v  0.909154550 -0.6576117921 -0.0078121968  0.259966453
#> v -1.362959424 -1.1516580441  0.1321602324  0.086642197
#> v -0.526109035 -0.9412824980  0.4425542265  0.701098015
#> v  1.463216788 -1.2270357371 -0.2214386462  0.882625342
#> v  0.521983414  0.0386469118  0.1033824181  0.003561641
#> v  2.050962711  0.5840034881 -0.1046903936 -0.542147880
#> v  0.317291974  0.1667156540 -1.6722424480 -0.028146024
#> v  0.400408709  1.0721119854  0.1266028756  0.045580705
#> v -0.061363860 -1.0764435907 -0.4167112917 -0.011272568
#> v  0.536224550 -1.3549772917  0.0342219194  0.240524068
#> v -2.529418708 -0.8185311433 -0.9294239444 -0.585987433
#> v -1.082493323  0.2058043597  0.6801214677  0.380506485
#> v -0.188792554  0.3127564538  0.1310483284  0.066479203
#> v -0.604344854  0.0452466495 -0.6760144393 -0.912723560
#> v -0.755682984 -0.2881793527 -0.0695078201 -0.554427026
#> v  0.524148520 -0.5917005292  0.3404292632  0.772159941
#> v  1.200353132 -0.9080713269 -1.0230624444 -0.324030846
#> v  0.408044953  0.8723428206  0.8118046022 -0.975328032