Generating spiral sample

n-2 points are drawn from a normal distribution with mean=0, sd=1, the points in the final two direction are sampled along a spiral by samping the angle from a normal distribution with mean=0, sd=2*pi (absolute values are used to fix the orientation of the spiral).

spiralData(n, p)

Arguments

n

sample dimensionality

p

number of sample points to generate

Value

sample points in matrix format

Examples

spiralData(4, 100)
#>           [,1]         [,2]         [,3]         [,4]
#> v -0.478355050  0.568878744  1.135767153  0.854174038
#> v -0.003301496  0.416918015 -0.392390315  0.065013723
#> v -0.062925119 -1.636219349  0.409896976  0.726343166
#> v -0.408686195 -0.094117059 -0.385657824  0.079640232
#> v -0.828195377  0.396813229  0.008801339  0.253455724
#> v  0.627437586 -0.805206472  1.196010446 -0.544067836
#> v  0.717082011  1.792486559  0.576158134  0.563234593
#> v  1.609947684  1.732799496 -0.144969024  0.889816437
#> v  0.424007988 -0.137778131 -0.388436750 -0.274595557
#> v -0.722073272  0.072762483 -1.011407874  0.170876115
#> v -0.316206077 -1.414721408  1.342873747 -0.116908715
#> v  0.705734028 -0.790427631  0.030706348 -0.575748891
#> v  2.078375621  1.919134900  0.131161377  0.067380986
#> v  0.267751527  0.436290736  0.215214776  0.832747047
#> v -1.253055561  1.765028630  0.107940352  0.168495947
#> v  0.898874217 -0.704995332  0.564528403  0.578145726
#> v -1.974322214 -1.244772830  0.733581652  0.145920322
#> v  0.579097050 -1.037890010 -1.001161577 -0.408241124
#> v  0.101959622  2.324238302 -0.716283035  1.387383316
#> v -1.293149956 -0.045018855  0.074608671  0.210104630
#> v  0.480623395  2.017342464  0.285276581  0.801996252
#> v  0.656575520  1.026297846 -0.423415803 -0.054010665
#> v  1.120615649  0.399897655  0.715203961 -0.069748844
#> v -0.502562184  0.987148440  0.530898960  1.378231500
#> v -0.165042212 -0.686040800  0.645152605 -0.248894676
#> v -0.164042843 -1.302339410 -0.091968643 -0.546899421
#> v  1.390088740  0.681840626 -0.386403374  0.078097177
#> v  0.420921089 -0.800331265 -0.405837275  0.029484492
#> v  0.539004516  1.435171017 -0.019658005  0.263786653
#> v -1.418623721 -0.513792720  0.081738894 -0.579515044
#> v  1.403682339 -0.015804912  0.735011476  0.108330478
#> v -2.116392160  0.150371437 -0.423604366 -0.055871104
#> v -0.905454844 -0.749986903  0.135361206 -0.578597526
#> v -0.559681870 -1.390851305 -0.056513143  0.271768387
#> v -0.480284997 -0.967621014  0.212578495 -1.198405224
#> v -0.069089037 -0.426188249 -0.013358404  1.514539772
#> v -0.914170952 -0.800524582  0.011907840 -0.573192382
#> v  0.660649173  1.727695639  0.120653857 -0.579341093
#> v -0.781818406 -0.125797549  0.734513038  0.070531296
#> v  0.294676272 -0.296458790  0.270561313 -0.554135407
#> v  1.691546077 -1.476545705  0.618469716 -0.292485561
#> v -0.613832093  0.617598091  0.301702433 -0.543805060
#> v -1.133375927 -0.156378319  0.010966905  0.252500998
#> v -1.129264413 -0.062192485 -0.404706412  0.032919186
#> v -0.054636953 -0.198624209 -0.970015940  0.292588006
#> v  0.198434624 -1.391001690  0.054686543  1.509103970
#> v -0.231792440 -0.686551360 -0.401102466  0.043223083
#> v -2.243508266 -0.381078571  0.128949824  0.115202082
#> v  0.892492885  1.808520401  0.682884373  0.373540927
#> v  1.452055280  2.556923589  0.694598066  0.341515199
#> v -1.178914989  1.075417132  0.126248614  0.044429020
#> v -0.234343833  0.950759734  0.130794019  0.064608055
#> v  0.663527998 -0.816198742 -0.080644360  0.274077828
#> v  1.346357662 -0.280015436  0.128800945  0.053867261
#> v -0.132833831  0.629361770 -0.317551091 -0.389958146
#> v -0.104018599 -1.388668827 -0.345928650  0.144016643
#> v  0.315862087  0.194610535 -0.371534624  0.106014158
#> v  0.812534989  0.275042593  0.103701857  0.003917462
#> v  2.010186412  0.313808823  0.358852573 -0.519825762
#> v -0.134641045  1.470832200 -1.014333210  0.159833253
#> v  0.204499683 -0.341768421  0.688112862 -1.052478424
#> v -0.205909535  1.502231278  0.012372570  0.251867538
#> v  0.053455720 -0.125146310  0.003554784  0.255664601
#> v -0.626612578  0.682877534 -0.397845612 -0.251101308
#> v -0.814967320 -0.345798987  0.127059612  0.047128024
#> v -0.566766181 -0.042837457 -0.178326169  0.887488885
#> v  0.967582242  1.029048853  0.732521193  1.264331104
#> v -0.303348223  0.179268039 -0.894919827  0.439143322
#> v -0.619946595 -0.131842450 -0.029507263  0.266479027
#> v -1.456778518  0.234340520  0.665746735 -0.209269313
#> v  1.898439401 -1.042770546 -0.471114933  0.811127341
#> v  1.764724810  0.506063430 -0.286858605 -0.423771199
#> v  0.667508839  1.372876146 -0.392392316 -0.265165606
#> v  0.809307437 -0.929351716  0.695263056 -0.137694465
#> v -0.563483049  1.888488707  0.128762851  0.116089482
#> v -0.049778439  0.412530895 -0.166062427  0.265320004
#> v -0.219463937 -0.279921977  0.086409306  0.197858236
#> v -1.131041190  0.599068143 -0.178099994  0.261979568
#> v  0.328683628 -0.002001793 -0.701878162 -0.889069045
#> v  0.830144580  1.311300782 -0.084526838 -0.549513393
#> v -0.388345065 -1.042408354 -0.035594589 -0.563811597
#> v -0.379227989  0.855040406  0.618549692 -0.292365280
#> v -0.511420306  0.575606727  0.359337371  0.760660471
#> v  0.873922726  0.752654962 -0.335794525 -0.366494722
#> v  0.489051671  0.493466172  1.355687435  0.211828267
#> v -0.115346633  0.402961199 -0.448321347  0.820944079
#> v  0.850032275  1.105681923 -0.291651665  0.200513450
#> v -0.123141511  0.448847720  1.273095491  0.581131888
#> v -1.249536842 -0.219531175  0.733346872  0.149480694
#> v -1.582876585  0.084967784  0.209898267  0.834780558
#> v -0.023462572  0.084693918 -0.138235762  0.271027122
#> v  0.505923564  0.415272046  0.369864885  0.753955463
#> v  0.033445330 -0.679873381  0.707449019 -0.099261153
#> v  1.772477385  0.277933293 -0.399376298  0.047856404
#> v -1.791060602  0.530663702  0.571268563  0.569592452
#> v  0.040226151 -1.187967396  0.011111736  1.512941457
#> v  1.553161749  0.323021948  0.113133611  0.016289685
#> v -0.249427874 -2.172319120 -0.502901402  0.796210125
#> v -0.980485754 -1.792780676 -0.424098701 -0.061208846
#> v -1.591328479  0.806263133 -0.213969585 -0.484508726