Generating spiral sample

n-2 points are drawn from a normal distribution with mean=0, sd=1, the points in the final two direction are sampled along a spiral by samping the angle from a normal distribution with mean=0, sd=2*pi (absolute values are used to fix the orientation of the spiral).

Usage

spiralData(n, p)

Arguments

n

sample dimensionality

p

number of sample points to generate

Value

sample points in matrix format

Examples

spiralData(4, 100)
#>           [,1]         [,2]         [,3]         [,4]
#> v  0.895414770 -0.083443880  0.043714283  0.879345122
#> v  0.071710106  0.381153132 -0.110503754 -0.539855101
#> v -1.208123013  1.070491316 -0.958322124  0.319788089
#> v -0.355679369 -0.607559598  1.257445230 -0.416559948
#> v  1.698005341 -0.738201192  0.689241018 -0.154281265
#> v  0.242499160 -0.815162137 -0.392795408 -0.264170926
#> v -0.800868011 -0.624816542 -1.119438979  1.131557781
#> v -0.715843977 -1.270220440 -0.027219800  0.888045276
#> v -0.257891448 -3.679007212 -0.364455165  0.117578399
#> v -1.347365506 -0.219303683 -0.687680838 -0.902235814
#> v  1.300838235 -0.081009412 -0.406327633 -0.225895436
#> v -0.760981191  0.598869689 -0.162441574 -0.515749577
#> v -0.798239439 -1.854334763 -0.109341699 -0.540319602
#> v -0.131468357 -0.275069062  0.157950704  0.852571635
#> v -0.288595190  1.239788854  0.126778885  0.124314499
#> v -2.094046724 -1.119598880  0.996111385 -0.816481457
#> v  0.340152799 -0.071719307  0.589029431 -0.333357754
#> v  1.824705023 -0.355623825  0.775033615  1.234538191
#> v -0.848446432  1.851398173  0.080247943  0.204504477
#> v -0.436746841 -0.406965380 -0.348961565  0.855700739
#> v -1.094729165 -0.373391575  0.503665694 -0.423963642
#> v -0.598744490 -0.687464324  0.246711514 -0.560814894
#> v  1.915922545 -0.091923142  0.192228008 -0.572222765
#> v  0.056954758 -0.332564397  0.552406464 -0.376477272
#> v  0.711169746 -0.041989329  0.120937931  0.142029728
#> v -0.555164755  0.288005112 -0.030746685  0.266787882
#> v  1.018037114  0.211094863 -0.066874629  0.273029573
#> v  0.365928319 -1.396460628  0.113142974  0.016304110
#> v -0.910517697  0.386489849  0.114113712  0.017828309
#> v -0.804496233 -0.166540862  0.425986281 -0.482414999
#> v -0.536234042 -0.185539382  0.026861009  0.244675172
#> v -0.294871345  0.707342896 -0.028410444  0.266200200
#> v  0.569362231 -1.043773591 -0.791177363  0.577352105
#> v -1.093359385 -2.964031913  0.132119227  0.088473296
#> v  1.897274942  1.249561489 -0.388118996  0.074477765
#> v  0.780582140 -2.012543530  0.132049273  0.090448695
#> v  0.509920643 -0.725138620  0.115757679  0.153933931
#> v -0.419696344  0.659925163  0.183047069 -0.573630239
#> v -0.664609851  0.521602156 -0.404486599  0.033574780
#> v  0.093885864 -0.752119179 -0.264443349  0.220899038
#> v -0.812580829 -1.155860853 -0.035334848 -0.563874611
#> v  0.600409023  0.101221234  0.049228439 -0.577646708
#> v  0.341127654  0.683741185  0.124766999  0.039994443
#> v -0.291531797 -1.388678100 -0.199827160  0.254569238
#> v  0.030633075 -0.249335058  0.572436605 -0.353818256
#> v -0.036688552 -0.130230355  0.112531653  0.015372801
#> v -1.594664359 -2.434916235 -0.259710322 -0.449127592
#> v -0.135360703 -2.523981771  0.458277695 -0.460329937
#> v  0.908689684  1.631233186 -0.308115023 -0.401030198
#> v -0.115632837 -0.890795496 -0.380409315 -0.292154608
#> v  0.910630913 -0.029664633 -0.292590668  0.199730655
#> v -0.063772035  1.468654127 -0.720050242 -0.871535124
#> v  0.255570223  0.024299091  0.096555459  0.185443476
#> v  1.589806608  0.948880032 -0.425307322 -0.080333429
#> v -1.324115477 -1.603107928 -0.239724646 -1.154456315
#> v  0.252915526 -0.203602189 -0.310025593  0.184107557
#> v -0.086544252 -0.259015582 -0.060437756 -0.557165417
#> v -0.715312421  0.731111456 -0.174537223 -0.509126430
#> v  0.937127434 -1.015356369  0.092451394  0.190713819
#> v  1.103117171  0.292140255  1.150254435 -0.620908577
#> v  0.764505697 -0.377514384 -0.425540173 -0.107292102
#> v  1.388600824 -0.424895077 -0.218770505  0.246591555
#> v -1.419396315  0.692795974  1.222390685  0.700030771
#> v -0.003858885 -0.871392046 -0.020470609 -0.567262560
#> v -0.259155608 -1.473526736 -0.570458423  0.759434143
#> v -0.216946324 -0.471020663 -0.371547822  0.105991693
#> v -0.971953648  0.844885266 -0.390368998  0.069577180
#> v -0.533805581  0.926368708 -0.298061239  0.195055285
#> v -1.575998241  0.958101944  0.114916525  0.019133391
#> v  0.278616729  0.939134624 -0.320654075 -0.386168751
#> v  0.556841494 -0.642014236 -0.360555690 -0.329050209
#> v  0.749577500 -0.548980021  0.094568099  0.188041707
#> v  2.076907498  0.131077511 -0.252131289 -0.455575701
#> v  1.305320690  0.007253685 -0.361808624 -0.326936638
#> v  1.668426736 -1.316765063  0.713899372 -0.075094832
#> v  0.955897056  0.578969677  0.110357954  0.012225510
#> v -1.864518981  0.078841083 -0.177483260  0.262163763
#> v -0.145291622 -0.463711175  0.654189514  0.438084298
#> v  0.024409177 -0.890439725  0.087780892  0.196294836
#> v -0.516747178  0.398999144  0.727707153 -0.004354291
#> v -0.319483584 -0.639762156 -0.812840087  0.552413770
#> v -0.436491649  0.303567608  0.543189171  0.603743057
#> v  0.215418682  0.209324028 -0.016473852  0.262823681
#> v  0.558972939 -0.521812892  0.455020990 -0.462690907
#> v -0.323846498  0.436659883  0.724607262 -0.024017502
#> v  2.692515039 -1.861033697  0.605022164  0.522752936
#> v  0.737202133  1.518832777  0.115695657 -0.579508114
#> v  0.725283091  0.648490725 -0.380948457  0.845889720
#> v -0.110312948  1.444280315 -0.008081274  0.260060561
#> v -0.012173077 -1.588928185 -0.419780918 -0.027382351
#> v -0.354420614  0.420952386  0.473647577  0.674479720
#> v -0.445451455 -0.086206632  0.649772201 -0.240535100
#> v -0.677462097 -0.824404203  0.126039314  0.126975854
#> v -1.995084156 -1.239298910  0.091606825  0.191754112
#> v -0.816230818  1.020711732 -1.041451598 -0.209640097
#> v  1.154428245 -0.051759966 -0.055568076  0.271632737
#> v -0.303187060  0.156827019 -0.181757678  0.260857897
#> v  0.996510324 -1.901515646 -0.385304257  0.080365900
#> v  1.322971167 -1.826342700 -0.190943591 -0.499470061
#> v -1.276119338  0.983569482  1.274793843 -0.372800058