Generating spiral sample

n-2 points are drawn from a normal distribution with mean=0, sd=1, the points in the final two direction are sampled along a spiral by samping the angle from a normal distribution with mean=0, sd=2*pi (absolute values are used to fix the orientation of the spiral).

Usage

spiral_data(n, p)

Arguments

n

number of sample points to generate

p

sample dimensionality

Value

sample points in matrix format

Examples

spiral_data(100, 4)
#> # A tibble: 100 × 4
#>         V1      V2     V3       V4
#>      <dbl>   <dbl>  <dbl>    <dbl>
#>  1  1.72   -0.453   0.105  0.173  
#>  2  1.63    0.158   0.118  0.0253 
#>  3  0.0820  1.06   -0.280 -1.79   
#>  4 -0.117   0.796   0.167 -0.576  
#>  5  0.125  -0.346  -0.407  0.0265 
#>  6  0.0369  0.944  -0.426 -0.102  
#>  7 -0.201   0.0305  0.124  0.0385 
#>  8  1.59    1.10   -0.387 -1.10   
#>  9 -1.48    0.0200  0.729  0.00821
#> 10 -1.20    0.633   0.104  0.175  
#> # ℹ 90 more rows