Estimating squint angle of 2-d structure in high-d dataset under selected index.

We estimate the squint angle by interpolating from a random starting plane towards the optimal view until the index value of the selected index function is above the selected cutoff. Since this depends on the direction, this is repeated with n randomly selected planes giving a distribution representative of the squint angle.

squintAngleEstimate(
  data,
  indexF,
  cutoff,
  structurePlane,
  n = 100,
  stepSize = 0.01
)

Arguments

data

Input data

indexF

Index function

cutoff

Threshold index value above which we assume the structure to be visible

structurePlane

Plane defining the optimal view

n

Number of random starting planes (default = 100)

stepSize

Interpolation step size fixing the accuracy (default = 0.01)

Value

numeric vector containing all squint angle estimates

Examples

data <- spiralData(4, 50)
indexF <- scagIndex("stringy")
cutoff <- 0.7
structurePlane <- basisMatrix(3,4,4)
squintAngleEstimate(data, indexF, cutoff, structurePlane, n=1)
#> [1] 0.7908864