Time each index evaluation for projections in the tour path.

Index evaluation timing may depend on the data distribution, we evaluate the computing time for a set of different projections to get an overview of the distribution of computing times.

timeSequence(d, t, idx, pmax)

Arguments

d
Input data in matrix format

t
List of projection matrices (e.g. interpolated tour path)

idx
Index function

pmax
Maximum number of projections to evaluate (cut t if longer than pmax)

Value

numeric vector containing all distances

Examples

# \donttest{
d <- spiralData(4, 1000)
t <- purrr::rerun(10, tourr::basis_random(4))
idx <- scagIndex("Skinny")
timeSequence(d, t, idx, 10)

#>       t  i
#> 1  0.04  1
#> 2  0.05  2
#> 3  0.05  3
#> 4  0.03  4
#> 5  0.07  5
#> 6  0.03  6
#> 7  0.06  7
#> 8  0.03  8
#> 9  0.03  9
#> 10 0.03 10

# }